fix: socket server setup hiiv
