+++ apps/L1G/tinder/socket-server/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "socket-server",
+  "version": "1.0.0",
+  "type": "module",
+  "main": "./dist/index.js",
+"scripts": {
+  "dev": "node --loader ts-node/esm src/server.ts",
+  "build": "tsc",
+  "start": "node dist/server.js"
+}
+  ,
+  "dependencies": {
+    "express": "^4.18.2",
+    "socket.io": "^4.7.2"
+  },
+  "devDependencies": {
+    "typescript": "^5.4.0",
+    "ts-node": "^10.9.1",
+    "@types/node": "^20.11.0",
+    "@types/express": "^4.17.21"
+  }
+}
